Refactor to move codecs to separate packages
